refactor: my enrolled course card
