+++ src/components/DashboardComponents/SubComponents/CourseEnrolledCard.js
@@ -7,23 +7,46 @@ import {
   CardContent,
   Typography,
   LinearProgress,
+  Box,
+  Chip,
 } from "@material-ui/core";
+import { useHistory } from "react-router";
 
 function CourseEnrolledCard() {
   const classes = useStyles();
+  const history = useHistory();
+
   return (
-    <Card className={classes.card}>
-      <CardHeader
-        className={classes.cardMedia}
-        style={{
-          backgroundImage: `url(https://d2slcw3kip6qmk.cloudfront.net/marketing/techblog/how-to-plan-a-programming-competition-header@2x.png)`,
-          backgroundSize: "cover",
-        }}
-      />
+    <Card
+      className={classes.card}
+      onClick={() =>
+        history.push(`/course/full-stack-webdevelopment/lecture/${125154558}#overview`)
+      }
+    >
       <CardActionArea>
+        <CardHeader
+          className={classes.cardMedia}
+          style={{
+            backgroundImage: `url(https://d2slcw3kip6qmk.cloudfront.net/marketing/techblog/how-to-plan-a-programming-competition-header@2x.png)`,
+            backgroundSize: "cover",
+          }}
+        />
         <CardContent className={classes.cardContent}>
-          <LinearProgress variant="determinate" value={45} />
-          <Typography className={classes.title}>Data Structures and Algorithms</Typography>
+          <Box classes={classes.tagSection}>
+            <Chip variant="outlined" size="small" className={classes.tag} label="Android" />
+            <Chip size="small" className={classes.price} label="₹1200" />
+          </Box>
+          <Box mt={2}>
+            <Typography>12/65 lessons Completed</Typography>
+          </Box>
+          <Box mt={2}>
+            <LinearProgress variant="determinate" value={40} />
+          </Box>
+          <Box className={classes.cardActions}>
+            <Typography variant="h6" className={classes.title}>
+              Full stack Web application Development Course By Code for Cause
+            </Typography>
+          </Box>
         </CardContent>
       </CardActionArea>
     </Card>
@@ -53,11 +76,20 @@ const useStyles = makeStyles((theme) => ({
   cardMedia: {
     height: 150,
   },
-  cardContent: {
-    // height: 120,
-  },
   title: {
     marginTop: theme.spacing(1),
+  },
+  tag: {
+    border: " 1px solid #3740A1",
+    borderRadius: "5px",
+    background: "#fff",
+    marginRight: theme.spacing(2),
+    color: theme.palette.primary.main,
+  },
+  price: {
+    background: theme.palette.primary.main,
+    borderRadius: "5px",
+    color: "#fff",
   },
 }));
